+++ src/solver/SequentialGreedy.cpp
@@ -13,7 +13,7 @@ void SequentialGreedy::solve(Timer &timer, long long iteration) {
 
     // initialize unvisited nodes, cost, tour
     list<Node> unvisited_nodes(this->graph->getNodes().begin(), this->graph->getNodes().end());
-    float cost = 0.0;
+    double cost = 0.;
     vector<Node> tour;
     tour.reserve(this->graph->getNodes().size());
